pulling out flightplan data into it's own data structure
POS and #M1BPOS messages can have the same information embedded in it. Will update in a future PR
